backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
pillow==10.0.1
opencv-python==4.8.1.78
scikit-learn==1.3.0
sqlite3
python-dotenv==1.0.0
werkzeug==2.3.7
gunicorn==21.2.0